Separate spgrep.representation
