+++ tests/test_irreps.py
@@ -17,18 +17,14 @@ from spgrep.irreps import (
     is_equivalent_irrep,
 )
 from spgrep.pointgroup import pg_dataset
-from spgrep.representation import (
-    check_spacegroup_representation,
-    get_character,
-    is_representation,
-    is_unitary,
-)
+from spgrep.rep.representation import get_character, is_representation, is_unitary
 from spgrep.symmetry.group import (
     check_cocycle_condition,
     get_cayley_table,
     get_factor_system_from_little_group,
     get_little_group,
 )
+from spgrep.symmetry.representation import check_spacegroup_representation
 from spgrep.transform import transform_symmetry_and_kpoint, unique_primitive_symmetry
 from spgrep.utils import NDArrayComplex, get_symmetry_from_hall_number
