+++ tests/test_spinor.py
@@ -7,11 +7,7 @@ from spgrep.core import (
     get_spacegroup_spinor_irreps_from_primitive_symmetry,
 )
 from spgrep.rep.group import get_identity_index
-from spgrep.representation import (
-    check_spacegroup_representation,
-    is_representation,
-    is_unitary,
-)
+from spgrep.rep.representation import is_representation, is_unitary
 from spgrep.spinor import (
     enumerate_spinor_small_representations,
     get_rotation_angle_and_axis,
@@ -22,6 +18,7 @@ from spgrep.symmetry.group import (
     get_cayley_table,
     get_little_group,
 )
+from spgrep.symmetry.representation import check_spacegroup_representation
 
 
 @pytest.mark.parametrize(
